refactor: cleanup imports and fix tests
