+++ blitzdb/__init__.py
@@ -1,5 +1,6 @@
-from .document import Document
 from .backends.file import Backend as FileBackend
+from .document import Document
+
 try:
     from .backends.mongo import Backend as MongoBackend
 except ImportError:
@@ -9,5 +10,5 @@ try:
     from .backends.sql import Backend as SqlBackend
 except ImportError:
     pass
-    
+
 __version__ = '0.2.12'
