+++ blitzdb/backends/base.py
@@ -1,9 +1,9 @@
 import abc
-import inspect
 import copy
-import six
-
+import inspect
 import logging
+
+import six
 
 logger = logging.getLogger(__name__)
