+++ blitzdb/backends/file/__init__.py
@@ -1,4 +1,4 @@
 from blitzdb.backends.file.backend import Backend
-from blitzdb.backends.file.store import Store, TransactionalStore
 from blitzdb.backends.file.index import Index, TransactionalIndex, NonUnique
 from blitzdb.backends.file.queryset import QuerySet
+from blitzdb.backends.file.store import Store, TransactionalStore
