+++ blitzdb/backends/file/index.py
@@ -1,12 +1,11 @@
 """File backend index."""
 import copy
-
 from collections import defaultdict
 
 from blitzdb.backends.base import NotInTransaction
-from blitzdb.backends.file.serializers import PickleSerializer as Serializer
-from blitzdb.backends.file.utils import JsonEncoder
 from blitzdb.backends.file.queryset import QuerySet
+from blitzdb.backends.file.serializers import PickleSerializer as Serializer
+
 
 class NonUnique(BaseException):
     """Index uniqueness constraint violated"""
