+++ blitzdb/backends/file/utils.py
@@ -1,9 +1,9 @@
-import json
 import datetime
+import json
 
 
 class JsonEncoder(json.JSONEncoder):
-    
+
     def default(self, obj):
         if isinstance(obj, set):
             return list(obj)
