+++ blitzdb/tests/file/test_unique_index.py
@@ -1,11 +1,10 @@
 from __future__ import absolute_import
 
 import pytest
-import tempfile
 
-from ..fixtures import *
-from ...backends.file import NonUnique
 from blitzdb.tests.helpers.movie_data import Actor
+from ...backends.file import NonUnique
+
 
 def test_nonunique_file_backend_index(file_backend):
     file_backend.create_index(Actor, fields={'yob': 1})
@@ -23,5 +22,6 @@ def test_unique_file_backend_index(file_backend):
     file_backend.save(actor1)
     actor2 = Actor({'num_films': 1})
     file_backend.save(actor2)
+
     with pytest.raises(NonUnique):
         file_backend.commit()
