+++ blitzdb/tests/sql/test_basics.py
@@ -1,9 +1,7 @@
 import pytest
-import pprint
 
-from ..helpers.movie_data import Movie,Actor,Director
+from ..helpers.movie_data import Movie, Actor, Director
 
-from .fixtures import backend
 
 def test_multiple_joins(backend):
