+++ blitzdb/tests/sql/test_eager_loading.py
@@ -1,10 +1,8 @@
- # -*- coding: utf-8 -*-
-import pytest
+# -*- coding: utf-8 -*-
 
-from ..helpers.movie_data import Movie,Actor,Director
-
-from .fixtures import backend
 from blitzdb.backends.sql.relations import ManyToManyProxy
+from ..helpers.movie_data import Movie, Actor, Director
+
 
 def prepare_data(backend):
