+++ blitzdb/tests/sql/test_enums.py
@@ -1,11 +1,9 @@
 import pytest
-import pprint
-
-from ..helpers.movie_data import Movie,Actor,Director,Document
-
-from .fixtures import backend
 from sqlalchemy.exc import IntegrityError
+
 from blitzdb.fields import EnumField
+from ..helpers.movie_data import Document
+
 
 class MyActor(Document):
