+++ blitzdb/tests/sql/test_one_to_many.py
@@ -1,11 +1,8 @@
 import pytest
-import pprint
 
-from ..helpers.movie_data import Movie,Actor,Director
-
-from .fixtures import backend
-from blitzdb.backends.sql.relations import ManyToManyProxy
 from blitzdb.backends.sql.queryset import QuerySet
+from ..helpers.movie_data import Movie, Actor, Director
+
 
 def test_basics(backend):
