+++ blitzdb/tests/test_array_queries.py
@@ -1,8 +1,8 @@
-import pytest
-from .helpers.movie_data import Movie,Actor,Director
-from .fixtures import *
+
 from blitzdb.backends.file import Backend as FileBackend
-    
+from .helpers.movie_data import Director, Movie, Actor
+
+
 def test_array_queries(backend):
 
     francis_coppola = Director({'name' : 'Francis Coppola'})
@@ -23,22 +23,19 @@ def test_array_queries(backend):
     backend.commit()
 
     result = backend.filter(Actor,{'movies' : the_godfather})
-
     assert len(result) == 2
     assert marlon_brando in result
     assert al_pacino in result
 
     result = backend.filter(Actor,{'movies' : {'$all' : [the_godfather,apocalypse_now]}})
-
     assert len(result) == 1
     assert marlon_brando in result
 
     result = backend.filter(Actor,{'movies.title' : {'$all' : ['The Godfather','Apocalypse Now']}})
-
     assert len(result) == 1
     assert marlon_brando in result
 
-    if not isinstance(backend,FileBackend):
+    if not isinstance(backend, FileBackend):
 
         #$elemMatch queries are currently not supported by the file backend.
 
@@ -54,32 +51,25 @@ def test_array_queries(backend):
         assert marlon_brando in result
 
     result = backend.filter(Actor,{'movies.title' : 'The Godfather'})
-
     assert len(result) == 2
     assert marlon_brando in result
     assert al_pacino in result
 
-
     result = backend.filter(Actor,{'movies' : {'$in' : [the_godfather,apocalypse_now]}})
-
     assert len(result) == 2
     assert marlon_brando in result
     assert al_pacino in result
 
     result = backend.filter(Actor,{'movies.title' : {'$in' : ['The Godfather','Apocalypse Now']}})
-
     assert len(result) == 2
     assert marlon_brando in result
     assert al_pacino in result
 
     result = backend.filter(Actor,{'$or' : [{'movies.title' : 'The Godfather'},{'movies.title' : 'Apocalypse Now'}]})
-
     assert len(result) == 2
     assert marlon_brando in result
     assert al_pacino in result
 
-
     result = backend.filter(Movie,{'director' : francis_coppola})
-
     assert len(result) == 1
     assert the_godfather in result
