+++ blitzdb/tests/test_dbref_includes.py
@@ -1,8 +1,7 @@
 from __future__ import absolute_import
 
-from .fixtures import *
-
 from blitzdb.tests.helpers.movie_data import Actor, Movie
+
 
 def test_nested_value(backend):
