+++ blitzdb/tests/test_get.py
@@ -1,8 +1,9 @@
 from __future__ import absolute_import
 
-from .fixtures import *
+import pytest
 
-from blitzdb.tests.helpers.movie_data import Actor, Director, Movie
+from blitzdb.tests.helpers.movie_data import Actor
+
 
 def test_delete(backend):
