+++ blitzdb/tests/test_transactions.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import
 
+import pytest
 
-from .fixtures import *
 from blitzdb.tests.helpers.movie_data import Movie
 
 
@@ -25,14 +25,14 @@ def test_rollback_and_commit(transactional_backend, small_transactional_test_dat
 
     (movies, actors, directors) = small_transactional_test_data
 
-    transactional_backend.begin()    
+    transactional_backend.begin()
     transactional_backend.filter(Movie, {}).delete()
     transactional_backend.rollback()
 
     assert len(movies) == len(transactional_backend.filter(Movie, {}))
     assert sorted(movies, key=lambda x: x.pk) == sorted(transactional_backend.filter(Movie, {}), key=lambda x: x.pk)
 
-    transactional_backend.begin()    
+    transactional_backend.begin()
     transactional_backend.filter(Movie, {}).delete()
     transactional_backend.commit()
